VarDiff working

+++ src/Configuration/ClusterConfig.cs
@@ -62,27 +62,27 @@ namespace MiningForce.Configuration
 		/// <summary>
 		/// Minimum difficulty
 		/// </summary>
-		public int MinDiff { get; set; }
+		public double MinDiff { get; set; }
 
 		/// <summary>
 		/// Network difficulty will be used if it is lower than this
 		/// </summary>
-		public int MaxDiff { get; set; }
+		public double MaxDiff { get; set; }
 
 		/// <summary>
 		/// Try to get 1 share per this many seconds
 		/// </summary>
-		public int TargetTime { get; set; }
+		public double TargetTime { get; set; }
 
 		/// <summary>
 		/// Check to see if we should retarget every this many seconds
 		/// </summary>
-		public int RetargetTime { get; set; }
+		public double RetargetTime { get; set; }
 
 		/// <summary>
 		/// Allow time to very this % from target without retargeting
 		/// </summary>
-		public int VariancePercent { get; set; }
+		public double VariancePercent { get; set; }
 	}
 
 	public class BanningConfig
